+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-07-16 15:29:12
-// Commit: fix(state): prevent double tap duplicate action in state
+// Updated at: 2026-07-16 16:37:54
+// Commit: feat(state): add bi-directional event bus handler for state
 
 class stateComponentState {
-  final String commitRef = '2895';
-  final String lastMessage = 'fix(state): prevent double tap duplicate action in state';
-  final DateTime timestamp = DateTime.parse('2026-07-16T15:29:12+05:00');
+  final String commitRef = '2897';
+  final String lastMessage = 'feat(state): add bi-directional event bus handler for state';
+  final DateTime timestamp = DateTime.parse('2026-07-16T16:37:54+05:00');
 }